Made clones of the original Text box, add course button and calculate Gp button. There are now a total of 5 instances for these elements placed side by side with each other.
Due to issues with vscode and the identification of wx packages on linux, a makeshift version of the
required binaries was added to the project directory in the 'usr/local/include/wx-3.1/wx' folder.
Still nothing seems to have been solved or corrected and I cannot build the source code from vscode.
I still have to use the terminal to do that.